refactor: Pasta separada para types
