+++ typescript-api/src/components/ListProducts.vue
@@ -26,7 +26,7 @@
 </template>
 <script lang="ts">
 import { Component, Emit, Prop, Vue } from "vue-property-decorator";
-import { IProduct } from "./types";
+import { IProduct } from "../types";
 
 @Component
 export default class ListProducts extends Vue {
